requirements.txt
colorama
discord-py-interactions
huggingface_hub